Rename chart builder tab component. Improve auto labeling. Add chart names for saving
