--- frontend/src/Monitor/Visualizations/Custom/Util.js
@@ -1,1 +1,0 @@
-function selectData()
